energy added

+++ src/methods/explicit_adams.cpp
@@ -1,9 +1,9 @@
 #include "explicit_adams.hpp"
 
-void explicit_adams::razgonka(double h, explicit_rk *rk, int size, double *x, std::function<double*(int, double*, void*)> f, void *data) {
-	adams_dx = new double[a_size * size];
-	double *local = new double[size];
-	double *dx;
+void explicit_adams::razgonka(long double h, explicit_rk *rk, int size, long double *x, std::function<long double*(int, long double*, void*)> f, void *data) {
+	adams_dx = new long double[a_size * size];
+	long double *local = new long double[size];
+	long double *dx;
 	for (int i = 0; i < a_size; i++) {
 		rk->rk_step(h, size, x, f, data);
 		for (int idx = 0; idx < size; idx++) {
@@ -18,7 +18,7 @@ void explicit_adams::razgonka(double h, explicit_rk *rk, int size, double *x, st
 	delete[] local;
 }
 
-void explicit_adams::adams_step(double h, int size, double *x, std::function<double*(int, double*, void*)> f, void * data) {
+void explicit_adams::adams_step(long double h, int size, long double *x, std::function<long double*(int, long double*, void*)> f, void * data) {
 	for (int ki = 0; ki < a_size; ki++) {
 		for (int idx = 0; idx < size; idx++) {
 			x[idx] += h * adams_a[ki] * adams_dx[ki * size + idx];
@@ -27,11 +27,11 @@ void explicit_adams::adams_step(double h, int size, double *x, std::function<dou
 	for (int i = 0; i < (a_size-1) * size; i++) {
 		adams_dx[i] = adams_dx[i+size];
 	}
-	double *local = new double[size];
+	long double *local = new long double[size];
 	for (int i = 0; i < size; i++) {
 		local[i] = x[i];
 	}
-	double *dx;
+	long double *dx;
 	dx = f(size, local, data);
 	for (int idx = 0; idx < size; idx++) {
 		adams_dx[(a_size-1) * size + idx] = dx[idx];
