+++ src/methods/explicit_adams.hpp
@@ -1,4 +1,4 @@
-#ifndef EXPLICIT_ADAMS
+	#ifndef EXPLICIT_ADAMS
 #define EXPLICIT_ADAMS
 
 #include <functional>
@@ -7,14 +7,14 @@
 class explicit_adams {
 public:
 	int a_size;
-	double *adams_a;
-	double *adams_dx;
+	long double *adams_a;
+	long double *adams_dx;
 
-	void adams_step(double h, int size, double *x, std::function<double*(int, double*, void*)> f, void * data);
+	void adams_step(long double h, int size, long double *x, std::function<long double*(int, long double*, void*)> f, void * data);
 
-	void razgonka(double h, explicit_rk *rk, int size, double *x, std::function<double*(int, double*, void*)> f, void *data);
+	void razgonka(long double h, explicit_rk *rk, int size, long double *x, std::function<long double*(int, long double*, void*)> f, void *data);
 
-	explicit_adams(int a_size, double *adams_a)
+	explicit_adams(int a_size, long double *adams_a)
 	:a_size(a_size), adams_a(adams_a) {};
 };
